TkrReasonsTool: found a double that should have been an int... don't think it was causing errors TkrCalibAlg: vstart and vend now print out correctly

+++ TkrUtil/src/TkrReasonsTool.cxx
@@ -58,7 +58,9 @@ namespace {
 
     double _xTower;
     double _yTower;
-    double _tower;
+    int   
+
+ _tower;
     int    _layer;
     int    _view;
     int    _tray;
